Remove usage of celo_attestation_stats throughout application (#352)
* Remove CeloAttestationStats and references.

* Modify usage of celo_attestation_stats to reference attestation properties on celo_account.

* Minor refactoring.

* Add rebuild_attestation_stats query.

* Move celo event definitions into dedicated module.

* Move attestation stats logic to explorer app.

* Test attestation stats update.

* Mix format.

* Fix credo.

* Mix format again aah.

* Fix validator augmentation.

* Remove commented code.
