+++ apps/block_scout_web/lib/block_scout_web/templates/address_celo/index.html.eex
@@ -64,7 +64,7 @@
       <dt class="col-sm-3 text-muted">Attestations requested</dt>
        <dd class="col-sm-9">
        <span data-toggle="tooltip" data-placement="top" title="" data-original-title="How many attestations have been requested from this account">
-       <%=  @address.celo_attestation_stats.requested %>
+       <%=  @address.celo_account.attestations_requested %>
        </span>
        </dd>
       </dl>
@@ -72,7 +72,7 @@
       <dt class="col-sm-3 text-muted">Attestations fulfilled</dt>
        <dd class="col-sm-9">
        <span data-toggle="tooltip" data-placement="top" title="" data-original-title="How many attestations have been fulfilled by this account">
-       <%= @address.celo_attestation_stats.fulfilled %></span>
+       <%= @address.celo_account.attestations_fulfilled %></span>
        </dd>
       </dl>
       <dl class="row">
