+++ apps/explorer/lib/explorer/chain/celo_signers.ex
@@ -7,7 +7,7 @@ defmodule Explorer.Chain.CeloSigners do
 
   use Explorer.Schema
 
-  alias Explorer.Chain.{Address, CeloAccount, CeloAttestationStats, CeloValidator, Hash}
+  alias Explorer.Chain.{Address, CeloAccount, CeloValidator, Hash}
 
   @typedoc """
   * `address` - address of the validator.
@@ -26,19 +26,6 @@ defmodule Explorer.Chain.CeloSigners do
   @required_attrs ~w(
           address signer
       )a
-
-  # Signer change events
-  @validator_signer_authorized "0x16e382723fb40543364faf68863212ba253a099607bf6d3a5b47e50a8bf94943"
-  @vote_signer_authorized "0xaab5f8a189373aaa290f42ae65ea5d7971b732366ca5bf66556e76263944af28"
-  @attestation_signer_authorized "0x9dfbc5a621c3e2d0d83beee687a17dfc796bbce2118793e5e254409bb265ca0b"
-
-  # Events for updating account
-  def signer_events,
-    do: [
-      @validator_signer_authorized,
-      @vote_signer_authorized,
-      @attestation_signer_authorized
-    ]
 
   schema "celo_signers" do
     belongs_to(
@@ -59,7 +46,6 @@ defmodule Explorer.Chain.CeloSigners do
 
     has_one(:celo_account, CeloAccount, foreign_key: :address, references: :address)
     has_one(:celo_validator, CeloValidator, foreign_key: :address, references: :address)
-    has_one(:celo_attestation_stats, CeloAttestationStats, foreign_key: :address_hash, references: :address)
 
     timestamps(null: false, type: :utc_datetime_usec)
   end
